Moved functions from GOES download to more logical util submodules
